requirements.txt
Flask==2.0.2
requests==2.26.0
python-owasp-zap-v2.4==0.0.16
gunicorn==20.1.0
werkzeug==2.2.2
